wip: save/load streams
