+++ ConnectorRevit/ConnectorRevit/ConnectorRevit.projitems
@@ -37,12 +37,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)RevitVersionHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Revit\ErrorEater.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ExternalEventHandler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Revit\EventHandling.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Revit\FamilyLoadOption.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Storage\StreamStateManager2.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Storage\StreamStateManager.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsRevit2\ConnectorBindingsRevit2.Events.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsRevit2\ConnectorBindingsRevit2.Receive.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsRevit2\ConnectorBindingsRevit2.ClientOperations.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsRevit2\ConnectorBindingsRevit2.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsRevit.Selection.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)UI\ConnectorBindingsRevit.Receive.cs" />
