+++ ConnectorRevit/ConnectorRevit/Entry/App.cs
@@ -43,7 +43,7 @@ namespace Speckle.ConnectorRevit.Entry
       }
 
       //desktopui 2
-      var speckleButton2 = specklePanel.AddItem(new PushButtonData("Speckle 2 New Ui", "Revit Connector - New UI!", typeof(App).Assembly.Location, typeof(SpeckleRevitCommand2).FullName)) as PushButton;
+      var speckleButton2 = specklePanel.AddItem(new PushButtonData("Speckle 2 New Ui", "Revit Connector\nNew UI!", typeof(App).Assembly.Location, typeof(SpeckleRevitCommand2).FullName)) as PushButton;
 
       if (speckleButton2 != null)
       {
@@ -100,6 +100,8 @@ namespace Speckle.ConnectorRevit.Entry
 
       //pre build app, so that it's faster to open up
       SpeckleRevitCommand2.InitAvalonia();
+      SpeckleRevitCommand2.Bindings = new ConnectorBindingsRevit2(AppInstance);
+      SpeckleRevitCommand2.Bindings.RegisterAppEvents();
 
     }
